+++ src/modules/bgckmicrowave_mod.f90
@@ -1794,7 +1794,8 @@ contains
   subroutine mwbg_tovCheckAmsua(TOVERRST, clwThreshArr, errThreshAllsky, useStateDepSigmaObs, &
                                 IUTILST, KTERMER, ICANO, ZO, btClear, ZCOR, &
                                 ZOMP, ICHECK, KNO, KNT, KNOSAT, ISCNPOS, MGINTRP, MTINTRP, GLINTRP, ITERRAIN, SATZEN, &
-                                globMarq, IMARQ, ident, clwObs, clwFG, scatw, STNID, RESETQC, ZLAT)
+                                globMarq, IMARQ, ident, clwObs, clwFG, scatwObs, scatwFG, &
+                                STNID, RESETQC, ZLAT)
 
   
     !:Purpose:          Effectuer le controle de qualite des radiances tovs.
@@ -1852,7 +1853,8 @@ contains
     !                                                                     >0, rejet,
     real, allocatable, intent(out)         :: clwObs(:)                ! retrieved cloud liquid water from observation 
     real, allocatable, intent(out)         :: clwFG(:)                 ! retrieved cloud liquid water from background 
-    real, allocatable, intent(out)         :: scatw(:)                 ! scattering index over water
+    real, allocatable, intent(out)         :: scatwObs(:)              ! scattering index over water from observation
+    real, allocatable, intent(out)         :: scatwFG(:)               ! scattering index over water from background
 
     integer, allocatable, intent(out)      :: ident(:)                 !ATMS Information flag (ident) values (new BURP element 025174 in header)
     !                                                                   FOR AMSUA just fill with zeros
@@ -1940,7 +1942,8 @@ contains
     ! Allocation
     call utl_reAllocate(clwObs,  KNT)
     call utl_reAllocate(clwFG,   KNT)
-    call utl_reAllocate(scatw,   KNT)
+    call utl_reAllocate(scatwObs,KNT)
+    call utl_reAllocate(scatwFG, KNT)
 
     call utl_reAllocate(kchkprf, KNT)
     call utl_reAllocate(ident, KNT)
@@ -1973,7 +1976,7 @@ contains
 
     call grody (err, knt, tb23, tb31, tb50, tb53, tb89, tb23FG, tb31FG, &
                 satzen, zlat, ktermer, ice, tpw, clwObs, clwFG, &
-                rain, snow, scatl, scatw)   
+                rain, snow, scatl, scatwObs)   
 
     ! 10) test 10: RTTOV reject check (single)
     ! Rejected datum flag has bit #9 on.
@@ -2041,7 +2044,7 @@ contains
                                 ICLWREJ, KMARQ, ICHECK)
     ! 13) test 13: Grody scattering index check (partial)
     ! For Scattering Index > 9, reject AMSUA-A channels 1-6 and 15.
-    call amsuaTest13GrodyScatteringIndexCheck (KCANO, KNOSAT, KNO, KNT, STNID, scatw, KTERMER, ITERRAIN, &
+    call amsuaTest13GrodyScatteringIndexCheck (KCANO, KNOSAT, KNO, KNT, STNID, scatwObs, KTERMER, ITERRAIN, &
                                               MISGRODY, MXSCATREJ, ISCATREJ, KMARQ, ICHECK)
 
     ! 14) test 14: "Rogue check" for (O-P) Tb residuals out of range. (single/full)
@@ -3620,7 +3623,7 @@ contains
   subroutine mwbg_tovCheckAtms(TOVERRST, clwThreshArr, errThreshAllsky, useStateDepSigmaObs, &
                                IUTILST, zlat, zlon, ilq, itt, zenith, qcflag2, qcflag1, &
                                ICANO, ztb, biasCorr, ZOMP, ICHECK, KNO, KNT, KNOSAT, IDENT, &
-                               ISCNPOS, MTINTRP, globMarq, IMARQ, clwObs, clwFG, riwv, &
+                               ISCNPOS, MTINTRP, globMarq, IMARQ, clwObs, clwFG, scatwObs, scatwFG, &
                                STNID, RESETQC)
 
 
@@ -3667,7 +3670,8 @@ contains
     !                                                             (chech n2) par satellite, critere et par canal
     real, allocatable, intent(out)   :: clwObs(:)
     real, allocatable, intent(out)   :: clwFG(:)
-    real, allocatable, intent(out)   :: riwv(:)
+    real, allocatable, intent(out)   :: scatwObs(:)            ! scattering index over water from observation
+    real, allocatable, intent(out)   :: scatwFG(:)             ! scattering index over water from background
 
     !locals
     real                             :: PTBOMP(KNO,KNT)
@@ -3688,6 +3692,7 @@ contains
     real, allocatable                :: scatec(:)
     real, allocatable                :: scatbg(:)
     real, allocatable                :: SeaIce(:)
+    real, allocatable                :: riwv(:)
 
     integer, parameter               :: maxScanAngleAMSU = 96
     integer, parameter               :: MXSFCREJ   = 8
@@ -3832,8 +3837,9 @@ contains
 
     call mwbg_reviewAllCritforFinalFlagsAtms(KNT, KNO, lqc, grossrej, waterobs, &
                                              precipobs, clwObs, clwFG, scatec, scatbg, &
-                                             iwvreject, riwv, IMARQ, globMarq, zdi, ident, &
-                                             drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
+                                             scatwObs, scatwFG, iwvreject, riwv, IMARQ, &
+                                             globMarq, zdi, ident, drycnt, landcnt, &
+                                             rejcnt, iwvcnt, pcpcnt, flgcnt, &
                                              MXCLWREJ, chanIgnoreInAllskyGenCoeff, icano)
 
     !###############################################################################
@@ -3959,7 +3965,7 @@ contains
   subroutine mwbg_tovCheckMwhs2(TOVERRST, clwThreshArr, errThreshAllsky, useStateDepSigmaObs, &
                                 IUTILST, zlat, zlon, ilq, itt, zenith, &
                                 ICANO, ztb, biasCorr, ZOMP, ICHECK, KNO, KNT, KNOSAT, IDENT, &
-                                ISCNPOS, MTINTRP, globMarq, IMARQ, clwObs, clwFG, riwv, &
+                                ISCNPOS, MTINTRP, globMarq, IMARQ, clwObs, clwFG, scatwObs, scatwFG, &
                                 STNID, RESETQC, modLSQ, lastHeader)
 
 
@@ -4007,7 +4013,8 @@ contains
     !                                                             (chech n2) par satellite, critere et par canal
     real, allocatable, intent(out)   :: clwObs(:)
     real, allocatable, intent(out)   :: clwFG(:)
-    real, allocatable, intent(out)   :: riwv(:)
+    real, allocatable, intent(out)   :: scatwObs(:)            ! scattering index over water from observation
+    real, allocatable, intent(out)   :: scatwFG(:)             ! scattering index over water from background
 
     !locals
     real                             :: PTBOMP(KNO,KNT)
@@ -4028,6 +4035,7 @@ contains
     real, allocatable                :: scatec(:)
     real, allocatable                :: scatbg(:)
     real, allocatable                :: SeaIce(:)
+    real, allocatable                :: riwv(:)
 
     integer, parameter               :: maxScanAngleAMSU = 98
     integer, parameter               :: MXSFCREJ   = 8
@@ -4166,8 +4174,9 @@ contains
 
     call mwbg_reviewAllCritforFinalFlagsMwhs2(KNT, KNO, lqc, grossrej, trn, waterobs, &
                                               precipobs, clwObs, clwFG, scatec, scatbg, &
-                                              iwvreject, riwv, IMARQ, globMarq, zdi, ident, &
-                                              allcnt, drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
+                                              scatwObs, scatwFG, iwvreject, riwv, IMARQ, &
+                                              globMarq, zdi, ident, allcnt, drycnt, landcnt, &
+                                              rejcnt, iwvcnt, pcpcnt, flgcnt, &
                                               MXCLWREJ, chanIgnoreInAllskyGenCoeff, icano)
 
     !###############################################################################
@@ -6506,8 +6515,9 @@ contains
   !--------------------------------------------------------------------------
   subroutine mwbg_reviewAllCritforFinalFlagsAtms(nt, nval, lqc, grossrej, waterobs, &
                                                  precipobs, clwObs, clwFG, scatec, scatbg, &
-                                                 iwvreject, riwv, IMARQ, globMarq, zdi, ident, &
-                                                 drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
+                                                 scatwObs, scatwFG, iwvreject, riwv, IMARQ, &
+                                                 globMarq, zdi, ident, drycnt, landcnt, &
+                                                 rejcnt, iwvcnt, pcpcnt, flgcnt, &
                                                  MXCLWREJ, chanIgnoreInAllskyGenCoeff, icano)
 
     !:Purpose:                   Review all the checks previously made to determine which obs are to be accepted
@@ -6525,13 +6535,15 @@ contains
     real, intent(inout)                        :: clwFG(:)
     real, intent(in)                           :: scatec(:)
     real, intent(in)                           :: scatbg(:)
+    real, allocatable, intent(out)             :: scatwObs(:)
+    real, allocatable, intent(out)             :: scatwFG(:)    
     logical, intent(in)                        :: grossrej(:)
     logical, intent(in)                        :: waterobs(:)
     logical, intent(in)                        :: iwvreject(:)
     logical, intent(in)                        :: precipobs(:)
     integer, intent(inout)                     :: ident(:)
     real, intent(in)                           :: zdi(:)
-    real, intent(inout)                        :: riwv(:)
+    real, intent(in)                           :: riwv(:)
     integer, intent(inout)                     :: IMARQ(:)
     integer, intent(inout)                     :: globMarq(:)
     integer, intent(inout)                     :: drycnt
@@ -6552,6 +6564,8 @@ contains
 
     ! Allocation
     call utl_reAllocate(lflagchn,nt, nval)
+    call utl_reAllocate(scatwObs,nt)
+    call utl_reAllocate(scatwFG, nt)
 
     lflagchn(:,:) = lqc(:,:)  ! initialize with flags set in mwbg_firstQcCheckAtms
     do kk = 1, nt
@@ -6651,12 +6665,13 @@ contains
       if ( ANY(lflagchn(kk,:)) ) flgcnt = flgcnt + 1
     end do
 
-    ! RESET riwv array to ECMWF scattering index for output to BURP file
-    riwv(:) = scatec(:)
-    ! Set missing clwObs and riwv to BURP missing value (mwbg_realMissing)
+    ! RESET scatwObs array to ECMWF scattering index for output to BURP file
+    scatwObs(:) = scatec(:)
+    ! Set missing clwObs and scatwObs to BURP missing value (mwbg_realMissing)
     where (clwObs == -99. ) clwObs = mwbg_realMissing
     where (clwObs == -99. ) clwFG = mwbg_realMissing
-    where (riwv == -99. ) riwv = mwbg_realMissing
+    where (scatwObs == -99. ) scatwObs = mwbg_realMissing
+    scatwFG(:) = mwbg_realMissing
 
     ! Modify data flag values (set bit 7) for rejected data
     ! In all-sky mode, turn on bit=23 for channels in chanIgnoreInAllskyGenCoeff(:)
@@ -6694,8 +6709,9 @@ contains
   !--------------------------------------------------------------------------
   subroutine mwbg_reviewAllCritforFinalFlagsMwhs2(nt, nval, lqc, grossrej, trn, waterobs, &
                                                   precipobs, clwObs, clwFG, scatec, scatbg, &
-                                                  iwvreject, riwv, IMARQ, globMarq, zdi, ident, &
-                                                  allcnt, drycnt, landcnt, rejcnt, iwvcnt, pcpcnt, flgcnt, &
+                                                  scatwObs, scatwFG, iwvreject, riwv, IMARQ, &
+                                                  globMarq, zdi, ident, allcnt, drycnt, landcnt, &
+                                                  rejcnt, iwvcnt, pcpcnt, flgcnt, &
                                                   MXCLWREJ, chanIgnoreInAllskyGenCoeff, icano)
 
     !:Purpose:                   Review all the checks previously made to determine which obs are to be accepted
@@ -6713,13 +6729,15 @@ contains
     real, intent(inout)                        :: clwFG(:)
     real, intent(in)                           :: scatec(:)
     real, intent(in)                           :: scatbg(:)
+    real, allocatable, intent(out)             :: scatwObs(:)
+    real, allocatable, intent(out)             :: scatwFG(:)
     logical, intent(in)                        :: grossrej(:)
     logical, intent(in)                        :: waterobs(:)
     logical, intent(in)                        :: iwvreject(:)
     logical, intent(in)                        :: precipobs(:)
     integer, intent(inout)                     :: ident(:)
     real, intent(in)                           :: zdi(:)
-    real, intent(inout)                        :: riwv(:)
+    real, intent(in)                           :: riwv(:)
     integer, intent(inout)                     :: IMARQ(:)
     integer, intent(inout)                     :: globMarq(:)
     integer, intent(inout)                     :: allcnt
@@ -6743,6 +6761,8 @@ contains
 
     ! Allocation
     call utl_reAllocate(lflagchn,nt, nval)
+    call utl_reAllocate(scatwObs,nt)
+    call utl_reAllocate(scatwFG, nt)
 
     lflagchn(:,:) = lqc(:,:)  ! initialize with flags set in mwbg_firstQcCheckMwhs2
     do kk = 1, nt
@@ -6852,12 +6872,13 @@ contains
       if ( ANY(lflagchn(kk,:)) ) flgcnt = flgcnt + 1
     end do
 
-    ! RESET riwv array to Bennartz-Grody scattering index for output to BURP file
-    riwv(:) = scatbg(:)
-    ! Set missing clwObs and riwv to BURP missing value (mwbg_realMissing)
+    ! RESET scatwObs array to Bennartz-Grody scattering index for output to BURP file
+    scatwObs(:) = scatbg(:)
+    ! Set missing clwObs and scatwObs to BURP missing value (mwbg_realMissing)
     where (clwObs == -99. ) clwObs = mwbg_realMissing
     where (clwObs == -99. ) clwFG = mwbg_realMissing
-    where (riwv == -99. ) riwv = mwbg_realMissing
+    where (scatwObs == -99. ) scatwObs = mwbg_realMissing
+    scatwFG(:) = mwbg_realMissing
 
     ! Modify data flag values (set bit 7) for rejected data
     ! In all-sky mode, turn on bit=23 for channels in chanIgnoreInAllskyGenCoeff(:)
@@ -7269,8 +7290,9 @@ contains
                                 satScanPosition, modelInterpGroundIce, modelInterpTerrain,&
                                 modelInterpSeaIce, terrainTypeIndice, satZenithAngle,     &
                                 obsGlobalMarker, obsFlags, newInformationFlag, &
-                                cloudLiquidWaterPathObs, cloudLiquidWaterPathFG,      &
-                                atmScatteringIndexObs, burpFileSatId, RESETQC, obsLatitude)
+                                cloudLiquidWaterPathObs, cloudLiquidWaterPathFG, &
+                                atmScatteringIndexObs, atmScatteringIndexFG, &
+                                burpFileSatId, RESETQC, obsLatitude)
       else if (instName == 'AMSUB') then
         call mwbg_tovCheckAmsub(oer_toverrst, oer_cldPredThresh, oer_errThreshAllsky, oer_useStateDepSigmaObs, &
                                 oer_tovutil, landQualifierIndice,&
@@ -7291,7 +7313,7 @@ contains
                                newInformationFlag, satScanPosition,   &
                                modelInterpTerrain, obsGlobalMarker, obsFlags,            &
                                cloudLiquidWaterPathObs, cloudLiquidWaterPathFG, &
-                               atmScatteringIndexObs, burpFileSatId, RESETQC)
+                               atmScatteringIndexObs, atmScatteringIndexFG, burpFileSatId, RESETQC)
       else if (instName == 'MWHS2') then
         call mwbg_tovCheckMwhs2(oer_toverrst, oer_cldPredThresh, oer_errThreshAllsky, oer_useStateDepSigmaObs, &
                                 oer_tovutil, obsLatitude, obsLongitude, &
@@ -7301,7 +7323,8 @@ contains
                                 newInformationFlag, satScanPosition,   &
                                 modelInterpTerrain, obsGlobalMarker, obsFlags,            &
                                 cloudLiquidWaterPathObs, cloudLiquidWaterPathFG, &
-                                atmScatteringIndexObs, burpFileSatId, RESETQC, modLSQ, lastHeader)
+                                atmScatteringIndexObs, atmScatteringIndexFG, &
+                                burpFileSatId, RESETQC, modLSQ, lastHeader)
       else
         write(*,*) 'midas-bgckMW: instName = ', instName
         call utl_abort('midas-bgckMW: unknown instName')